Struct slog_json::JsonBuilder

pub struct JsonBuilder<W: Write> { /* fields omitted */ }

Json Drain builder

Create with Json::new.

Methods

impl<W> JsonBuilder<W> where
W: Write,

pub fn build(self) -> Json<W>

Build Json Drain

This consumes the builder.

pub fn set_newlines(self, enabled: bool) -> Self

Set writing a newline after every log record

pub fn set_pretty(self, enabled: bool) -> Self

Set whether or not pretty formatted logging should be used

pub fn add_key_value<T>(self, value: OwnedKV<T>) -> Self where
T: SendSyncRefUnwindSafeKV + 'static,

Add custom values to be printed with this formatter

pub fn add_default_keys(self) -> Self

Add default key-values:

ts - timestamp

level - record logging level name

msg - msg - formatted logging message